# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Calculator server."
  version: "1.0.0"
  title: "ApiCalculator"
  contact:
    email: "duenasr.juanc@javeriana.eedu.co"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "apiCalculator"
basePath: "/v1"
tags:
- name: "sum"
  description: "Returns the result of the sum of n numbers"
- name: "res"
  description: "Returns the result of the subtraction of two numbers"
- name: "mul"
  description: "Returns the result of the multiplication  of n numbers"
- name: "div"
  description: "Returns the result of the multiplication  of two numbers"
schemes:
- "https"
- "http"
paths:
  /sum/{num1}/..{numn}:
    get:
      tags:
      - "sum"
      summary: "Sum two numbers"
      description: "Returns the sum"
      operationId: "getSum"
      produces:
      - "application/json"
      parameters:
      - name: "num1"
        in: "path"
        description: "First Number to sum"
        required: true
        type: "integer"
        format: "int64"
      - name: "numn"
        in: "path"
        description: "N Number to sum"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Sum"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Operation not found"
  /res/{num1}/{num2}:
    get:
      tags:
      - "res"
      summary: "Substract two numbers"
      description: "Returns the subtraction "
      operationId: "getRes"
      produces:
      - "application/json"
      parameters:
      - name: "num1"
        in: "path"
        description: "First Number to substract"
        required: true
        type: "integer"
        format: "int64"
      - name: "num2"
        in: "path"
        description: "Second Number to substract"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Res"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Operation not found"
  /mul/{num1}/..{numn}:
    get:
      tags:
      - "mul"
      summary: "Multiply n numbers"
      description: "Returns the multiplication"
      operationId: "getMul"
      produces:
      - "application/json"
      parameters:
      - name: "num1"
        in: "path"
        description: "First Number to multiply"
        required: true
        type: "integer"
        format: "int64"
      - name: "numn"
        in: "path"
        description: "N Number to multiply"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Mul"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Operation not found"
  /div/{num1}/{num2}:
    get:
      tags:
      - "div"
      summary: "Divide two numbers"
      description: "Returns the division"
      operationId: "getDiv"
      produces:
      - "application/json"
      parameters:
      - name: "num1"
        in: "path"
        description: "First Number to divide"
        required: true
        type: "integer"
        format: "int64"
      - name: "num2"
        in: "path"
        description: "Second Number to divide"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Div"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Operation not found"
definitions:
  Sum:
    type: "object"
    properties:
      operation:
        type: "string"
        example: "sum"
      result:
        type: "integer"
        example:  4 
  Res:
    type: "object"
    properties:
      operation:
        type: "string"
        example: "res"
      result:
        type: "integer"
        example:  4 
  Mul:
    type: "object"
    properties:
      operation:
        type: "string"
        example: "mul"
      result:
        type: "integer"
        example:  4 
  Div:
    type: "object"
    properties:
      operation:
        type: "string"
        example: "div"
      result:
        type: "integer"
        example:  4     